Move numpy-dependent Python functions out of utils.py.
